+++ syncopy/tests/test_continuousdata.py
@@ -196,12 +196,14 @@ class TestAnalogData():
 
         # -- test with single array--
 
-        dummy = AnalogData(data=self.data)
+        chan_labels = [str(i) for i in range(self.nc)]
+        dummy = AnalogData(data=self.data, channel=chan_labels)
         assert dummy.dimord == AnalogData._defaultDimord
         assert dummy.channel.size == self.nc
         assert (dummy.sampleinfo == [0, self.ns]).min()
         assert dummy.trialinfo.shape == (1, 0)
         assert np.array_equal(dummy.data, self.data)
+        assert np.all(dummy.channel == chan_labels)
 
         # wrong shape for data-type
         with pytest.raises(SPYValueError):